remove dep on autograph/signer/xpi
The `go.mozilla.org/autograph/signer/xpi` dependency was forcing
autograph-edge to have a long, complicated build due to the PKCS#11
deps.

But the only thing we were using it for was a small JSON API object
whose definition is relatively static and easily copied. Better to have
both systems have separate copies to ensure backwards compatibility
happens, anyway.

This means we can speed up our builds by not needing clang or libltdl
installed. That work can be done in a follow-up PR.