correct deploy workflows by renaming generate target (#338)
In https://github.com/mozilla-services/autograph-edge/pull/337, we
reference `make generate` like in the autograph repo.

But here that was called `make version.json`. For consistency, and any
possible (though, unlikely) additions to what needs to be generate, we
rename the make target to `generate` and leave the deploy workflow
alone.